// src/contracts/generated/TaskStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskId } from "./TaskId.js";
import type { TaskState } from "./TaskState.js";

/**
 * 状态查询结果。
 */
export type TaskStatus = { 
/**
 * 任务 id。
 */
task_id: TaskId, 
/**
 * 当前状态。
 */
state: TaskState, 
/**
 * 当前阶段描述（执行中才有）。
 */
phase?: string, };